Better support different clock domains.
Prior to this patch, an arbitrator did not properly detect multiple
requests by the same process in the same cycle: if the same process A
requested at t0 and t1, and another process B requested at t1 just
before A, it would consider that A was requesting two times, which is
incorrect.

This patch properly remembers the last request cycle separately for
each requester process.